unit tests

+++ Restraunt.Data/Repositories/TableRepository.cs
@@ -22,33 +22,39 @@ namespace Restraunt.Data.Repositories
         }
         public async Task<bool> Create(TableDto? entity)
         {
-           var result = new Table { Id = entity.Id, Name = entity.Name, Link = entity.Link};
+            var result = new Table { Id = entity.Id, Name = entity.Name, Link = entity.Link };
 
             await _db.Tables.AddAsync(result);
-            
+
             return true;
         }
 
         public async Task<bool> Delete(Guid Id)
         {
             var entity = await _db.Tables.Where(t => t.Id == Id).FirstOrDefaultAsync();
-            _db.Tables.Remove(entity);
+            if (entity != null)
+            {
+                _db.Tables.Remove(entity);
+                return true;
+            }
+            else
+            {
+                return false;
+            }
 
-            
-            return true;
         }
 
         public async Task<Table> Get(Guid id)
         {
+            var table = await _db.Tables.Where(t => t.Id == id).Include(u => u.User).FirstOrDefaultAsync();
 
-            var table = await _db.Tables.Where(t => t.Id == id).Include(u=>u.User).FirstOrDefaultAsync();
 
             return table;
         }
 
         public async Task<IEnumerable<Table>> Select()
         {
-            return _db.Tables.Include(u=>u.User).ToList();
+            return _db.Tables.Include(u => u.User).ToList();
         }
 
         public async Task<Table> Edit(EditTableDto entity)
@@ -60,8 +66,10 @@ namespace Restraunt.Data.Repositories
                 table.Name = entity.Name;
 
                 _db.Update(table);
+
             }
             return table;
+
         }
 
         public async Task<Table> BindUserToTable(BindUserToTableDto model)
@@ -70,7 +78,7 @@ namespace Restraunt.Data.Repositories
 
             User? user = await _db.Users.FirstOrDefaultAsync(u => u.Id == model.UserId);
 
-            if(table != null)
+            if (table != null)
             {
                 table.User = user;
 
@@ -78,9 +86,11 @@ namespace Restraunt.Data.Repositories
                 _db.Update(table);
             }
             return table;
-            
+
         }
 
+
+      
 
     }
 }
